restructure role-assigment #29955
