+++ projects/safe/src/lib/components/role-summary/role-assignment/add-role-rule/add-role-rule.component.html
