+++ projects/safe/src/lib/components/role-summary/role-assignment/role-assignment.component.scss
@@ -1,0 +1,7 @@
+.header {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: top;
+}
