+++ projects/safe/src/lib/components/role-summary/role-assignment/role-assignment.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RoleAssignmentComponent } from './role-assignment.component';
+
+describe('RoleAssignmentComponent', () => {
+  let component: RoleAssignmentComponent;
+  let fixture: ComponentFixture<RoleAssignmentComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ RoleAssignmentComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(RoleAssignmentComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
