--- projects/safe/src/lib/components/role-summary/role-details/auto-role-assignment/auto-role-assignment.component.scss
@@ -1,16 +1,0 @@
-:host {
-  display: flex;
-  flex-direction: column;
-}
-
-safe-button {
-  align-self: end;
-}
-
-.capitalize::first-letter {
-  text-transform: uppercase;
-}
-
-table {
-  margin-top: 1em;
-}
