+++ projects/safe/src/lib/components/role-summary/role-details/role-details.component.html
@@ -1,12 +1,42 @@
-<safe-general-role-details
-  [application]="application"
-  [loading]="loading"
-  [role]="role"
-  (edit)="onUpdate($event)"
+<form [formGroup]="form">
+  <div class="form-group">
+    <mat-form-field appearance="outline">
+      <mat-label>
+        {{ 'models.role.title' | translate }}
+      </mat-label>
+      <input matInput formControlName="title" type="text" />
+    </mat-form-field>
+    <mat-form-field appearance="outline">
+      <mat-label>
+        {{ 'models.role.description' | translate }}
+      </mat-label>
+      <textarea
+        matInput
+        formControlName="description"
+        [cdkTextareaAutosize]="true"
+        [cdkAutosizeMinRows]="5"
+      ></textarea>
+    </mat-form-field>
+    <mat-form-field appearance="outline">
+      <mat-label>{{
+        'components.role.update.permissions' | translate
+      }}</mat-label>
+      <mat-select formControlName="permissions" multiple>
+        <mat-option
+          *ngFor="let permission of permissions"
+          [value]="permission.id"
+        >
+          {{ permission.type }}
+        </mat-option>
+      </mat-select>
+    </mat-form-field>
+  </div>
+</form>
+<safe-button
+  variant="primary"
+  category="secondary"
+  type="submit"
+  [disabled]="!form.valid"
+  (click)="onUpdate()"
+  >{{ 'common.saveChanges' | translate }}</safe-button
 >
-</safe-general-role-details>
-<safe-auto-role-assignment
-  [loading]="loading"
-  [application]="application"
-  [role]="role"
-></safe-auto-role-assignment>
