--- projects/safe/src/lib/components/role-summary/role-details/role-details.component.scss
@@ -1,5 +1,0 @@
-:host {
-  display: flex;
-  flex-direction: column;
-  gap: 2em;
-}
