+++ projects/safe/src/lib/components/role-summary/role-summary.component.html
@@ -13,6 +13,17 @@
       <safe-spinner *ngIf="loading"></safe-spinner>
     </ng-template>
   </mat-tab>
+  <mat-tab [label]="'components.role.summary.automaticAssignment.title' | translate">
+    <ng-template matTabContent>
+      <safe-role-assignment
+        *ngIf="role"
+        [role]="role"
+        (edit)="onEditRole($event)"
+        [loading]="loading"
+      ></safe-role-assignment>
+      <safe-spinner *ngIf="loading"></safe-spinner>
+    </ng-template>
+  </mat-tab>
   <mat-tab [label]="'common.resource.few' | translate">
     <ng-template matTabContent>
       <safe-role-resources *ngIf="role" [role]="role"></safe-role-resources>
