+++ projects/safe/src/lib/components/users/users.component.html
@@ -70,7 +70,7 @@
     >
       {{ 'components.users.invite.confirm' | translate }}
     </safe-button>
-    <safe-button
+    <!-- <safe-button
       *ngIf="advancedSettings"
       icon="settings"
       category="secondary"
@@ -78,7 +78,7 @@
       (click)="onAdvancedSettings()"
       [title]="'common.advancedSettings' | translate"
     >
-    </safe-button>
+    </safe-button> -->
   </div>
 </div>
 <div>
